+++ services/src/main/java/org/keycloak/protocol/oidc/mappers/HardcodedClaimBasedOnIdPAttributesMapper.java
@@ -32,7 +32,9 @@ public class HardcodedClaimBasedOnIdPAttributesMapper extends AbstractOIDCProtoc
         property.setName(ProtocolMapperUtils.CLAIM_VALUE);
         property.setLabel(ProtocolMapperUtils.CLAIM_VALUE_LABEL);
         property.setHelpText(ProtocolMapperUtils.CONDITIONAL_CLAIM_VALUE_HELP_TEXT);
-        property.setType(ProviderConfigProperty.STRING_TYPE);
+        property.setType(ProviderConfigProperty.MULTIVALUED_STRING_TYPE);
+        property.setStringify(Boolean.TRUE);
+        property.setDefaultValue("");
         configProperties.add(property);
 
         property = new ProviderConfigProperty();
@@ -46,7 +48,9 @@ public class HardcodedClaimBasedOnIdPAttributesMapper extends AbstractOIDCProtoc
         property.setName(ProtocolMapperUtils.IDP_ATTRIBUTE_VALUES);
         property.setLabel(ProtocolMapperUtils.IDP_ATTRIBUTE_VALUES_LABEL);
         property.setHelpText(ProtocolMapperUtils.IDP_ATTRIBUTE_VALUES_HELP_TEXT);
-        property.setType(ProviderConfigProperty.STRING_TYPE);
+        property.setType(ProviderConfigProperty.MULTIVALUED_STRING_TYPE);
+        property.setStringify(Boolean.TRUE);
+        property.setDefaultValue("");
         configProperties.add(property);
 
         OIDCAttributeMapperHelper.addAttributeConfig(configProperties, HardcodedClaimBasedOnAttributeMapper.class);
@@ -96,7 +100,7 @@ public class HardcodedClaimBasedOnIdPAttributesMapper extends AbstractOIDCProtoc
                     OIDCAttributeMapperHelper.mapClaim(token, mappingModel, attributeValues.size() == 1 ? attributeValues.get(0) : attributeValues);
                 }
             } catch (IOException e) {
-                logger.warn("problem executing HardcodedClaimBasedOnIdPAttributesMapper");
+                logger.warn("problem executing HardcodedAttributeBasedOnIdPAttributesMapper");
                 e.printStackTrace();
             }
         }